renaming the ccd config file
